requirements.txt
mediapipe>=0.10
opencv-python>=4.8
pyautogui>=0.9.54
pystray>=0.19
Pillow>=10.0
cryptography>=42.0
pynput>=1.7
plyer>=2.1
numpy>=1.24